# openapi.json
{
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "properties": {
          "knowledge_base_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Knowledge bases this question may draw on. Omit — or send an empty list — to search every base.",
            "title": "Knowledge Base Ids"
          },
          "query": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "required": [
          "query"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "ChatResponse": {
        "properties": {
          "intent": {
            "title": "Intent",
            "type": "string"
          },
          "is_valid": {
            "title": "Is Valid",
            "type": "boolean"
          },
          "iteration": {
            "title": "Iteration",
            "type": "integer"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "response": {
            "title": "Response",
            "type": "string"
          },
          "retrieved_count": {
            "title": "Retrieved Count",
            "type": "integer"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "sources": {
            "description": "Distinct documents behind this answer, most relevant first",
            "items": {
              "$ref": "#/components/schemas/SourceRef"
            },
            "title": "Sources",
            "type": "array"
          },
          "validation_reason": {
            "title": "Validation Reason",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "query",
          "intent",
          "response",
          "retrieved_count",
          "iteration",
          "is_valid",
          "validation_reason"
        ],
        "title": "ChatResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SessionSummary": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "number"
          },
          "message_count": {
            "title": "Message Count",
            "type": "integer"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "number"
          }
        },
        "required": [
          "session_id",
          "title",
          "created_at",
          "updated_at",
          "message_count"
        ],
        "title": "SessionSummary",
        "type": "object"
      },
      "SourceRef": {
        "properties": {
          "document": {
            "description": "File the passage came from",
            "title": "Document",
            "type": "string"
          },
          "knowledge_base_id": {
            "title": "Knowledge Base Id",
            "type": "string"
          },
          "knowledge_base_name": {
            "title": "Knowledge Base Name",
            "type": "string"
          }
        },
        "required": [
          "document",
          "knowledge_base_id",
          "knowledge_base_name"
        ],
        "title": "SourceRef",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "跨领域科研问答助手：多知识库 RAG 检索、引用溯源与 LangGraph 工作流",
    "title": "AI 科研助手",
    "version": "3.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/chat": {
      "post": {
        "description": "Run the graph and return a complete response.",
        "operationId": "chat_api_v1_chat_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Client-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Client-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/session/{session_id}": {
      "delete": {
        "operationId": "clear_session_api_v1_chat_session__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Client-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Client-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clear Session",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/session/{session_id}/history": {
      "get": {
        "operationId": "get_history_api_v1_chat_session__session_id__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Client-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Client-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get History",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/sessions": {
      "get": {
        "description": "List every session the calling token owns, newest first.\n\nSessions with no recorded meta (e.g. claimed before this feature shipped)\nare returned with a placeholder title so the UI can still show them; their\ntimestamp is the message-key creation time, which is best-effort.",
        "operationId": "list_sessions_api_v1_chat_sessions_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Client-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Client-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionSummary"
                  },
                  "title": "Response List Sessions Api V1 Chat Sessions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Sessions",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/stream": {
      "post": {
        "description": "Stream graph progress and model tokens as Server-Sent Events.",
        "operationId": "chat_stream_api_v1_chat_stream_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Client-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Client-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat Stream",
        "tags": [
          "Chat"
        ]
      }
    },
...
  }
}